requirements.txt
easyocr
mss
Pillow
pynput
openai
numpy
pygame
